requirements.txt
Flask==2.3.3
slack-bolt==1.18.0
openai>=1.50.0
gunicorn
python-dotenv
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0